Phase 7: Implement Deep AI Analysis and CodeBERT Integration Pipeline
- Added CodeBERT result model
- Implemented DeepAnalysisService
- Added IDeepAnalysisService interface
- Added ICodeBertService interface
- Implemented CodeBERT analysis pipeline
- Integrated Deep Analysis into ExecutionPipelineService
- Added Confidence scoring engine
- Updated ExecutionHistory entity
- Updated ExecutionResult entity
- Enhanced static analyzers (Go, Java, Python)
- Updated dependency injection for AI services
- Refactored execution pipeline for multi-layer AI analysis

This completes Phase 7 core AI intelligence architecture.

+++ CodeInterviewPro.Domain/Entities/ExecutionHistory.cs
@@ -16,7 +16,7 @@
 
         public int Score { get; set; }
 
-        public int AIScore { get; set; }
+        public double AIScore { get; set; }
 
         public string AIFeedback { get; set; }
